Physics Engine fixed - dv should not be conserved
Ok I finally realized the issue with the physics. Acceleration was
accidentally being conserved because I forgot to zero it out between
recalculations. Lol...well its a good thing because we now get smooth
and mathematically predictable orbits, which is really good.

I broke the camera with this update, but its on purpose. I've got a good
idea of what im doing next