You can now select and track objects in space
Just have to right click. You were right about how to do it; you take
the matrix and unproject it. XNA has a nice built in method for it too,
which is nice. I will continue to add in some game elements, and
hopefully we can link up the networking soon-ish.